pyproject.toml
[project]
name = "nusmt-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.2",
    "asyncpg>=0.31.0",
    "bcrypt<4.0",
    "fastapi[standard]>=0.128.0",
    "itsdangerous>=2.2.0",
    "passlib>=1.7.4",
    "pyjwt>=2.10.1",
    "sqladmin>=0.22.0",
    "sqlalchemy>=2.0.46",
]
